Los controladores para las multas y las citas hechos, yacada ventana tiene los controlaodres necesarios
